// src/types/mock-data.type.ts
import { Genre } from './genre.enum.js';

export type MockData = {
  titles: string[];
  descriptions: string[];
  publicationDates: Date[];
  genres: Genre[];
  years: number[];
  ratings: number[];
  previewVideoLinks: string[];
  videoLinks: string[];
  actors: string[];
  directors: string[];
  runTimes: number[];
  commentsCount: number[];
  names: string[];
  emails: string[];
  avatars: string[];
  passwords: string[];
  posterImages: string[];
  backgroundImages: string[];
  backgroundColors: string[];
};